chore: remove duplicate serde helpers
