+++ crates/rpc/rpc-types/src/eth/engine/payload.rs
@@ -447,7 +447,7 @@ pub struct OptimismPayloadAttributes {
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
-        deserialize_with = "crate::serde_helpers::u64_hex::u64_hex_opt::deserialize"
+        deserialize_with = "crate::serde_helpers::u64_hex_opt::deserialize"
     )]
     pub gas_limit: Option<u64>,
 }
